cdk/requirements.txt
aws-cdk-lib==2.140.0
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.aws-cloudfront-s3==2.56.0